Refactor cursor manager and commands
