+++ Cycloside/Services/RelayCommand.cs
@@ -1,0 +1,28 @@
+using System;
+using System.Windows.Input;
+
+namespace Cycloside.Services
+{
+    /// <summary>
+    /// A reusable command that relays its functionality to delegates.
+    /// </summary>
+    public class RelayCommand : ICommand
+    {
+        private readonly Action<object?> _execute;
+        private readonly Predicate<object?>? _canExecute;
+
+        public event EventHandler? CanExecuteChanged;
+
+        public RelayCommand(Action<object?> execute, Predicate<object?>? canExecute = null)
+        {
+            _execute = execute ?? throw new ArgumentNullException(nameof(execute));
+            _canExecute = canExecute;
+        }
+
+        public RelayCommand(Action execute) : this(_ => execute()) { }
+
+        public bool CanExecute(object? parameter) => _canExecute == null || _canExecute(parameter);
+
+        public void Execute(object? parameter) => _execute(parameter);
+    }
+}
